Bean Validation

+++ src/main/java/com/softcraft/sccustomerapi/services/ClientService.java
@@ -3,10 +3,14 @@ package com.softcraft.sccustomerapi.services;
 import com.softcraft.sccustomerapi.dto.ClientDTO;
 import com.softcraft.sccustomerapi.entities.Client;
 import com.softcraft.sccustomerapi.repositories.ClientRepository;
+import com.softcraft.sccustomerapi.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
@@ -19,10 +23,9 @@ public class ClientService {
 
     @Transactional(readOnly = true)
     public ClientDTO findById(Long id) {
-        Optional<Client> result = repository.findById(id);
-        Client client = result.get();
-        ClientDTO dto = new ClientDTO(client);
-        return dto;
+        Client client = repository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Cliente inexistente!"));
+        return new ClientDTO(client);
     }
 
     @Transactional(readOnly = true)
@@ -41,15 +44,23 @@ public class ClientService {
 
     @Transactional
     public ClientDTO update(Long id, ClientDTO dto) {
-        Client entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ClientDTO(entity);
+        try {
+            Client entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ClientDTO(entity);
+        }
+        catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Cliente inexistente!!");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Cliente inexistente!!!");
+        }
+            repository.deleteById(id);
     }
 
     private void copyDtoToEntity(ClientDTO dto, Client entity) {
